Refactor book borrowing system: - Removed update.php and replaced it with update_borrow.php for better clarity. - Added borrow.php as the main entry point for book borrowing management. - Created borrow_form.php for creating new borrow transactions. - Implemented borrowlist.php to display all borrow transactions with options to edit or delete. - Added create_borrow.php to handle the creation of new borrow transactions. - Introduced db.php for database connection management. - Developed delete_borrow.php for transaction deletion functionality. - Created functions.php for reusable validation and data fetching functions.
